requirements.txt
requests
python-whois